+++ packages/block-editor/src/private-apis.js
@@ -74,6 +74,9 @@ import {
 	useBlockElement,
 } from './components/block-list/use-block-props/use-block-refs';
 import { LinkPicker } from './components/link-picker';
+// Same component as the public LinkControl export. Exposed privately so that
+// `transformSuggestions` can be used without becoming supported public API.
+import PrivateLinkControl from './components/link-control';
 import useRemoteUrlData from './components/link-control/use-rich-url-data';
 import { PrivateBlockContext } from './components/block-list/private-block-context';
 import useListViewPanelState from './components/use-list-view-panel-state';
@@ -154,6 +157,7 @@ lock( privateApis, {
 	useBlockElement,
 	useBlockElementRef,
 	LinkPicker,
+	PrivateLinkControl,
 	useRemoteUrlData,
 	PrivateBlockContext,
 	useListViewPanelState,